Temporary fix for some warnings
